UI: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about shadowing variables
* Fixes warnings about unused assignments

+++ UI/auth-oauth.cpp
@@ -91,7 +91,7 @@ int OAuthLogin::exec()
 void OAuthLogin::urlChanged(const QString &url)
 {
 	std::string uri = get_token ? "access_token=" : "code=";
-	int code_idx = url.indexOf(uri.c_str());
+	int code_idx = (int)url.indexOf(uri.c_str());
 	if (code_idx == -1)
 		return;
 
@@ -100,7 +100,7 @@ void OAuthLogin::urlChanged(const QString &url)
 
 	code_idx += (int)uri.size();
 
-	int next_idx = url.indexOf("&", code_idx);
+	int next_idx = (int)url.indexOf("&", code_idx);
 	if (next_idx != -1)
 		code = url.mid(code_idx, next_idx - code_idx);
 	else
